Commit realizado pelo pycharm

+++ Exercicios/DESAFIO 07.py
@@ -1,4 +1,4 @@
-nota1 = int(input('Nota da primeira prova: '))
-nota2 = int(input('Nota da segunda prova: '))
+nota1 = float(input('Nota da primeira prova: '))
+nota2 = float(input('Nota da segunda prova: '))
 
 print('Nota final: {:.1f}'.format((nota1+nota2)/2))
